Move all Wi-Fi examples to `wifi` and convert to individual projects
