+++ examples/wifi/sniffer/Cargo.toml
@@ -1,0 +1,76 @@
+[package]
+name    = "sniffer"
+version = "0.0.0"
+edition = "2024"
+publish = false
+
+[dependencies]
+critical-section = "1.1.3"
+esp-alloc = { path = "../../../esp-alloc" }
+esp-backtrace = { path = "../../../esp-backtrace", features = [
+    "panic-handler",
+    "println",
+] }
+esp-bootloader-esp-idf = { path = "../../../esp-bootloader-esp-idf" }
+esp-hal = { path = "../../../esp-hal", features = ["log-04", "unstable"] }
+esp-println = { path = "../../../esp-println", features = ["log-04"] }
+esp-radio-preempt-baremetal = { path = "../../../esp-radio-preempt-baremetal", features = [
+    "log-04",
+] }
+esp-radio = { path = "../../../esp-radio", features = [
+    "log-04",
+    "sniffer",
+    "unstable",
+    "wifi",
+] }
+ieee80211 = { version = "0.4.0", default-features = false }
+
+[features]
+esp32 = [
+    "esp-backtrace/esp32",
+    "esp-bootloader-esp-idf/esp32",
+    "esp-hal/esp32",
+    "esp-radio-preempt-baremetal/esp32",
+    "esp-radio/esp32",
+]
+esp32c2 = [
+    "esp-backtrace/esp32c2",
+    "esp-bootloader-esp-idf/esp32c2",
+    "esp-hal/esp32c2",
+    "esp-radio-preempt-baremetal/esp32c2",
+    "esp-radio/esp32c2",
+]
+esp32c3 = [
+    "esp-backtrace/esp32c3",
+    "esp-bootloader-esp-idf/esp32c3",
+    "esp-hal/esp32c3",
+    "esp-radio-preempt-baremetal/esp32c3",
+    "esp-radio/esp32c3",
+]
+esp32c6 = [
+    "esp-backtrace/esp32c6",
+    "esp-bootloader-esp-idf/esp32c6",
+    "esp-hal/esp32c6",
+    "esp-radio-preempt-baremetal/esp32c6",
+    "esp-radio/esp32c6",
+]
+esp32s2 = [
+    "esp-backtrace/esp32s2",
+    "esp-bootloader-esp-idf/esp32s2",
+    "esp-hal/esp32s2",
+    "esp-radio-preempt-baremetal/esp32s2",
+    "esp-radio/esp32s2",
+]
+esp32s3 = [
+    "esp-backtrace/esp32s3",
+    "esp-bootloader-esp-idf/esp32s3",
+    "esp-hal/esp32s3",
+    "esp-radio-preempt-baremetal/esp32s3",
+    "esp-radio/esp32s3",
+]
+
+[profile.release]
+debug            = true
+debug-assertions = true
+lto              = "fat"
+codegen-units    = 1
