+++ examples/wifi/static_ip/src/main.rs
@@ -1,0 +1,214 @@
+//! Using a static IP example
+//!
+//! - set SSID and PASSWORD env variable
+//! - set STATIC_IP and GATEWAY_IP env variable (e.g. "192.168.2.191" / "192.168.2.1")
+//! - might be necessary to configure your WiFi access point accordingly
+//! - uses the given static IP
+//! - responds with some HTML content when connecting to port 8080
+
+#![no_std]
+#![no_main]
+
+use blocking_network_stack::Stack;
+use embedded_io::*;
+use esp_alloc as _;
+use esp_backtrace as _;
+use esp_hal::{
+    clock::CpuClock,
+    main,
+    rng::Rng,
+    time::{self, Duration},
+    timer::timg::TimerGroup,
+};
+use esp_println::{print, println};
+use esp_radio::wifi::{ClientConfiguration, Configuration};
+use smoltcp::iface::{SocketSet, SocketStorage};
+
+esp_bootloader_esp_idf::esp_app_desc!();
+
+const SSID: &str = env!("SSID");
+const PASSWORD: &str = env!("PASSWORD");
+const STATIC_IP: &str = env!("STATIC_IP");
+const GATEWAY_IP: &str = env!("GATEWAY_IP");
+
+#[main]
+fn main() -> ! {
+    esp_println::logger::init_logger_from_env();
+    let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
+    let peripherals = esp_hal::init(config);
+
+    esp_alloc::heap_allocator!(size: 72 * 1024);
+
+    let timg0 = TimerGroup::new(peripherals.TIMG0);
+    esp_radio_preempt_baremetal::init(timg0.timer0);
+
+    let esp_radio_ctrl = esp_radio::init().unwrap();
+
+    let (mut controller, interfaces) =
+        esp_radio::wifi::new(&esp_radio_ctrl, peripherals.WIFI).unwrap();
+
+    let mut device = interfaces.sta;
+    let iface = create_interface(&mut device);
+
+    controller
+        .set_power_saving(esp_radio::config::PowerSaveMode::None)
+        .unwrap();
+
+    let mut socket_set_entries: [SocketStorage; 3] = Default::default();
+    let socket_set = SocketSet::new(&mut socket_set_entries[..]);
+
+    let rng = Rng::new();
+    let now = || time::Instant::now().duration_since_epoch().as_millis();
+    let mut stack = Stack::new(iface, device, socket_set, now, rng.random());
+
+    let client_config = Configuration::Client(ClientConfiguration {
+        ssid: SSID.into(),
+        password: PASSWORD.into(),
+        ..Default::default()
+    });
+    let res = controller.set_configuration(&client_config);
+    println!("wifi_set_configuration returned {:?}", res);
+
+    controller.start().unwrap();
+    println!("is wifi started: {:?}", controller.is_started());
+
+    println!("Start Wifi Scan");
+    let res = controller.scan_n(10).unwrap();
+    for ap in res {
+        println!("{:?}", ap);
+    }
+
+    println!("{:?}", controller.capabilities());
+    println!("wifi_connect {:?}", controller.connect());
+
+    // wait to get connected
+    println!("Wait to get connected");
+    loop {
+        match controller.is_connected() {
+            Ok(true) => break,
+            Ok(false) => {}
+            Err(err) => {
+                println!("{:?}", err);
+                loop {}
+            }
+        }
+    }
+    println!("{:?}", controller.is_connected());
+
+    println!("Setting static IP {}", STATIC_IP);
+
+    stack
+        .set_iface_configuration(&blocking_network_stack::ipv4::Configuration::Client(
+            blocking_network_stack::ipv4::ClientConfiguration::Fixed(
+                blocking_network_stack::ipv4::ClientSettings {
+                    ip: blocking_network_stack::ipv4::Ipv4Addr::from(parse_ip(STATIC_IP)),
+                    subnet: blocking_network_stack::ipv4::Subnet {
+                        gateway: blocking_network_stack::ipv4::Ipv4Addr::from(parse_ip(GATEWAY_IP)),
+                        mask: blocking_network_stack::ipv4::Mask(24),
+                    },
+                    dns: None,
+                    secondary_dns: None,
+                },
+            ),
+        ))
+        .unwrap();
+
+    println!(
+        "Start busy loop on main. Point your browser to http://{}:8080/",
+        STATIC_IP
+    );
+
+    let mut rx_buffer = [0u8; 1536];
+    let mut tx_buffer = [0u8; 1536];
+    let mut socket = stack.get_socket(&mut rx_buffer, &mut tx_buffer);
+
+    socket.listen(8080).unwrap();
+
+    loop {
+        socket.work();
+
+        if !socket.is_open() {
+            socket.listen(8080).unwrap();
+        }
+
+        if socket.is_connected() {
+            println!("Connected");
+
+            let mut time_out = false;
+            let deadline = time::Instant::now() + Duration::from_secs(20);
+            let mut buffer = [0u8; 1024];
+            let mut pos = 0;
+            while let Ok(len) = socket.read(&mut buffer[pos..]) {
+                let to_print = unsafe { core::str::from_utf8_unchecked(&buffer[..(pos + len)]) };
+
+                if to_print.contains("\r\n\r\n") {
+                    print!("{}", to_print);
+                    println!();
+                    break;
+                }
+
+                pos += len;
+
+                if time::Instant::now() > deadline {
+                    println!("Timeout");
+                    time_out = true;
+                    break;
+                }
+            }
+
+            if !time_out {
+                socket.write_all(
+                    b"HTTP/1.0 200 OK\r\n\r\n\
+                    <html>\
+                        <body>\
+                            <h1>Hello Rust! Hello esp-radio!</h1>\
+                            <img src=\"https://rustacean.net/more-crabby-things/dancing-ferris.gif\"/>
+                        </body>\
+                    </html>\r\n\
+                    "
+                ).unwrap();
+
+                socket.flush().unwrap();
+            }
+
+            socket.close();
+
+            println!("Done\n");
+            println!();
+        }
+
+        let deadline = time::Instant::now() + Duration::from_secs(5);
+        while time::Instant::now() < deadline {
+            socket.work();
+        }
+    }
+}
+
+fn parse_ip(ip: &str) -> [u8; 4] {
+    let mut result = [0u8; 4];
+    for (idx, octet) in ip.split(".").into_iter().enumerate() {
+        result[idx] = u8::from_str_radix(octet, 10).unwrap();
+    }
+    result
+}
+
+// some smoltcp boilerplate
+fn timestamp() -> smoltcp::time::Instant {
+    smoltcp::time::Instant::from_micros(
+        esp_hal::time::Instant::now()
+            .duration_since_epoch()
+            .as_micros() as i64,
+    )
+}
+
+pub fn create_interface(device: &mut esp_radio::wifi::WifiDevice) -> smoltcp::iface::Interface {
+    // users could create multiple instances but since they only have one WifiDevice
+    // they probably can't do anything bad with that
+    smoltcp::iface::Interface::new(
+        smoltcp::iface::Config::new(smoltcp::wire::HardwareAddress::Ethernet(
+            smoltcp::wire::EthernetAddress::from_bytes(&device.mac_address()),
+        )),
+        device,
+        timestamp(),
+    )
+}
